dashboard: add sync indicator
A live-updated sync indicator, including snapshot download and RPC sync.
Looks crappy, but is functional. Needs some design attention.

There is currently a small gap of time after the snapshot where the
local block height lingers at zero, which distorts the time estimate for
the subsequent RPC sync; this probably should be tweaked in the API.